+++ src/SoundModules/InputNode.js
@@ -7,12 +7,12 @@ class InputNode {
         /** @type {undefined | Module} */
         this.input = undefined;
         this.owner = owner;
-        /** @returns {Promise<Float32Array>} */
+        /** @returns {Promise<Float64Array>} */
         this.getValues = async (recursion) => {
             // if(!this.input) throw new Error("requested getValues from nonconnected input");
             if (this.input)
                 return await this.input.getValues(recursion);
-            return new Float32Array(0);
+            return new Float64Array(0);
         };
         
         this.cacheObsolete = owner.cacheObsolete;
