+++ src/SoundModules/NaiveReverb.js
@@ -94,7 +94,7 @@ class NaiveReverb extends Module{
             tap1.reset();
 
 
-            this.cachedValues = new Float32Array(inputValues.length);
+            this.cachedValues = new Float64Array(inputValues.length);
             
             inputValues.forEach((value,sampleNumber)=>{
                 this.cachedValues[sampleNumber]=0;
