+++ src/SoundModules/Oscillator.js
@@ -87,7 +87,7 @@ class Oscillator extends Module{
         
         this.recalculate = async (recursion = 0) => {
             const lengthSamples = settings.length * sampleRate;
-            this.cachedValues = new Float32Array(lengthSamples);
+            this.cachedValues = new Float64Array(lengthSamples);
 
             operator.setShape(settings.shape);
             operator.setPhase(settings.phase);
