+++ src/SoundModules/Repeater.js
@@ -67,7 +67,7 @@ class Repeater extends Module {
         this.recalculate = async (recursion = 0) => {
 
             const lengthSamples = settings.length * sampleRate;
-            this.cachedValues = new Float32Array(lengthSamples);
+            this.cachedValues = new Float64Array(lengthSamples);
             
             sortPointsByTime();
             /** @returns {EnvelopePoint|false} */
